requirements/_base.txt
Django
South
git+https://github.com/nens/translations.git
django-autoslug
django-appconf
django-crispy-forms
